refactor: use u8 for 'dimension' index option
It falls into a small range of possible values so no need to
use u32. This saves us some otherwise wasted memory.

+++ sql-planner/src/ir/node.rs
@@ -1144,7 +1144,7 @@ pub struct CreateIndex {
     pub range_size: Option<u32>,
     pub run_count_per_level: Option<u32>,
     pub run_size_ratio: Option<Decimal>,
-    pub dimension: Option<u32>,
+    pub dimension: Option<u8>,
     pub distance: Option<RtreeIndexDistanceType>,
     pub hint: Option<bool>,
     pub wait_applied_globally: bool,
